teste de integração ok 100% cobertura

+++ src/main/java/com/renatinha/fleetmanagement/entities/LastTrajectory.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class LastTrajectory {
 
     @Id
-    private Long taxiId;
+    private int taxiId;
     private String plate;
     private String date;
     private Double longitude;
